# Make `SwiftRuntime.memory` constant property