… Plugins/PackageToJS/Templates/runtime.mjs
… Runtime/src/memory.ts