Also cache DataView instance as much as possible (until the underlying memory is grown)

```
==============================
   COMPARISON WITH BASELINE
==============================

Test                                                       | Status   | Baseline (ms)   | Current (ms)    | Change          | Change (%)
----------------------------------------------------------------------------------------------------------------------------------------
Call/JavaScript function call through Wasm import          | FASTER   | 125.69          | 117.39          | -8.30 ms        | -6.60%
Call/JavaScript function call through Wasm import with int | FASTER   | 75.97           | 70.69           | -5.28 ms        | -6.95%
Property access/Write Number                               | FASTER   | 484.17          | 411.26          | -72.91 ms       | -15.06%
Property access/Read Number                                | FASTER   | 621.07          | 432.51          | -188.56 ms      | -30.36%
Property access/Write String                               | FASTER   | 919.29          | 714.24          | -205.05 ms      | -22.31%
Property access/Read String                                | FASTER   | 1712.38         | 1239.51         | -472.87 ms      | -27.61%
```